Move disable patch into own component
