+++ src/components/PatchActionButtons/index.tsx
@@ -4,3 +4,4 @@ export { RestartPatch } from "./RestartPatch";
 export { SetPatchPriority } from "./SetPatchPriority";
 export { EnqueuePatch } from "./EnqueuePatch";
 export { AddNotification } from "./AddNotification";
+export { DisablePatch } from "./DisablePatch";
